strat: go all in on ncm::TitleId

+++ stratosphere/ams_mitm/source/ns_mitm/nsmitm_web_service.cpp
@@ -29,7 +29,7 @@ Result NsWebMitmService::GetDocumentInterface(Out<std::shared_ptr<NsDocumentServ
     R_TRY(nsGetDocumentInterfaceFwd(this->forward_service.get(), &doc));
 
     /* Set output interface. */
-    out_intf.SetValue(std::move(std::make_shared<NsDocumentService>(this->title_id, doc)));
+    out_intf.SetValue(std::move(std::make_shared<NsDocumentService>(static_cast<u64>(this->title_id), doc)));
     if (out_intf.IsDomain()) {
         out_intf.ChangeObjectId(doc.s.object_id);
     }
@@ -43,7 +43,7 @@ Result NsDocumentService::GetApplicationContentPath(OutBuffer<u8> out_path, u64 
 
 Result NsDocumentService::ResolveApplicationContentPath(u64 title_id, u8 storage_type) {
     /* Always succeed for web applet asking about HBL. */
-    if (Utils::IsWebAppletTid(this->title_id) && Utils::IsHblTid(title_id)) {
+    if (Utils::IsWebAppletTid(static_cast<u64>(this->title_id)) && Utils::IsHblTid(title_id)) {
         nswebResolveApplicationContentPath(this->srv.get(), title_id, static_cast<FsStorageId>(storage_type));
         return ResultSuccess;
     }
